fix: remove Reward contract as a parent contract, and let the child contracts all implement these functions seperately, also adjusted the tests for RewardShares and RewardTokens to both test all the functions individually.
